Move weebsh files to their own folder
