+++ src/main/kotlin/xyz/eroto/bot/commands/weebsh/Awoo.kt
@@ -1,4 +1,4 @@
-package xyz.eroto.bot.commands
+package xyz.eroto.bot.commands.weebsh
 
 import xyz.eroto.bot.entities.cmd.variant.WolkCommand
 import xyz.eroto.bot.utils.WolkType
